♻️ CR056N blueprints: decompose project-local registry
