+++ .agentplane/tasks/202605290053-CR056N/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290053-CR056N`
+Title: Blueprint project-local decomposition
+Canonical task record: `.agentplane/tasks/202605290053-CR056N/README.md`
+
+## Summary
+
+Blueprint project-local decomposition
+
+Decompose packages/agentplane/src/blueprints/project-local.ts into focused project-local blueprint modules while preserving behavior and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/blueprints/project-local.ts into focused project-local blueprint modules while preserving behavior and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "Blueprint project-local decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T00:54:59.111Z
+- Branch: task/202605290053-CR056N/blueprint-project-local-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
