+++ .agentplane/tasks/202605290053-CR056N/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 CR056N task: Blueprint project-local decomposition [202605290053-CR056N]
